refactor(server): modules
