+++ server/app.js
@@ -1,7 +1,10 @@
-const express = require('express')
+const express = require('express');
 
 const cors = require("cors");
+const mysql = require('mysql');
+const path = require('path');
 
+require('dotenv').config({path: path.resolve(__dirname+'/../.env')});
 
 
 
@@ -18,11 +21,11 @@ app.use(express.json());
 
 
 const signup = require('./routes/signup');
-const signin = require('./routes/signin');
+const signin = require('./routes/login');
 
 
 
-
+app.set("base",'/api')
