+++ server/db.js
@@ -1,5 +1,4 @@
-const mysql = require('mysql')
-require('dotenv').config()
+
